from_names to filter on players without ID
This change to from_names is useful when you do a lookup by ID but one
of the players you are looking for does not have an ID yet.  This can
happen for rookies.  You can do lookup by ID + lookup by name to get the
complete list of players you are after.